+++ src/lib.rs
@@ -8,6 +8,7 @@ use model_folder::ModelFolderState;
 use nutexb::NutexbFile;
 use nutexb_wgpu::TextureRenderer;
 use rfd::FileDialog;
+use serde::{Deserialize, Serialize};
 use ssbh_data::{matl_data::ParamId, prelude::*};
 use ssbh_wgpu::{
     swing::SwingPrc, ModelRenderOptions, RenderModel, RenderSettings, SharedRenderData,
@@ -62,31 +63,47 @@ impl<'a> TexturePainter<'a> {
 }
 
 // TODO: Separate input state and camera UI state?
-pub struct CameraInputState {
-    pub previous_cursor_position: PhysicalPosition<f64>,
-    pub is_mouse_left_clicked: bool,
-    pub is_mouse_right_clicked: bool,
-
-    pub translation: glam::Vec3,
-    pub rotation_radians: glam::Vec3,
-    pub fov_y_radians: f32,
+pub struct CameraState {
+    pub input: InputState,
+    pub values: CameraValues,
     pub anim_path: Option<PathBuf>,
 
     // TODO: Where to put this?
     pub mvp_matrix: glam::Mat4,
 }
 
-impl Default for CameraInputState {
+#[derive(Default)]
+pub struct InputState {
+    pub previous_cursor_position: PhysicalPosition<f64>,
+    pub is_mouse_left_clicked: bool,
+    pub is_mouse_right_clicked: bool,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(default)]
+pub struct CameraValues {
+    pub translation: glam::Vec3,
+    pub rotation_radians: glam::Vec3,
+    pub fov_y_radians: f32,
+}
+
+impl Default for CameraState {
     fn default() -> Self {
         Self {
-            previous_cursor_position: PhysicalPosition { x: 0.0, y: 0.0 },
-            is_mouse_left_clicked: false,
-            is_mouse_right_clicked: false,
+            input: InputState::default(),
+            values: CameraValues::default(),
+            anim_path: None,
+            mvp_matrix: glam::Mat4::IDENTITY,
+        }
+    }
+}
+
+impl Default for CameraValues {
+    fn default() -> Self {
+        Self {
             translation: glam::Vec3::new(0.0, -8.0, -60.0),
             rotation_radians: glam::Vec3::new(0.0, 0.0, 0.0),
             fov_y_radians: 30f32.to_radians(),
-            anim_path: None,
-            mvp_matrix: glam::Mat4::IDENTITY,
         }
     }
 }